requirements.txt
jsonschema
PyMuPDF
requests
python-dotenv
